+++ src/components/auth/Login.js
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom"
 import "./Login.css"
 
 export const Login = () => {
-    const [email, set] = useState("hpassfield7@netvibes.com")
+    const [email, set] = useState("fbaggins@shire.net")
     const navigate = useNavigate()
 
     const handleLogin = (e) => {
