+++ src/components/forms/edits/EditIngredients.js
@@ -1,6 +1,10 @@
-import Select from 'react-select';
+import { useState } from 'react';
+import { EditIngredientForm } from './EditIngredientForm';
+import { CustomIngredient } from '../CustomIngredient';
 
-export const EditIngredients = ({ingredientsToPost, allIngredients, ingredientToAdd, updateIngredientToAdd, updateIngredientsToPost, initialIngredients, ingredientsToDelete, updateIngredientsToDelete}) => {
+export const EditIngredients = ({ingredientsToPost, allIngredients, ingredientToAdd, updateIngredientToAdd, updateIngredientsToPost, initialIngredients, ingredientsToDelete, updateIngredientsToDelete, fetchList, setAllIngredients}) => {
+    // Set state variable for tracking if the create custom ingredient view should be shown
+    const [showCustom, setShowCustom] = useState(false)
     
     const handleAddIngredient = (event) => {
         event.preventDefault()
@@ -73,123 +77,27 @@ export const EditIngredients = ({ingredientsToPost, allIngredients, ingredientTo
     }
     
     return <>
-        <div className="addedIngredients">
-            { /*Initial ingredients already attached to the recipe card*/
-                initialIngredients.length > 0
-                && initialIngredients.map(initialIngredient => {
-                    const matchedIngredient = allIngredients.find(
-                        ingredient => ingredient.id === initialIngredient.ingredientId
-                    )
-                    return <div key={`displayInitIng--${initialIngredient.id}`}>
-                        { // If user has marked for delete, show 'undo' button
-                            markedForDeletion(initialIngredient)
-                                ? 
-                                    <div className="addedIngredientRow ingredientToDelete" key={`initialIngDetails1--${initialIngredient.ingredientId}`}>
-                                        <span className="flex-column1" key={`matchedIng--${initialIngredient.ingredientId}`}>{matchedIngredient?.name}</span>
-                                        <span className="flex-column2" key={`addedQuant--${initialIngredient.ingredientId}`}>{initialIngredient.quantity} {initialIngredient.quantityUnit}</span>
-                                        <span className="flex-column3" key={`removeIngredient--${initialIngredient.ingredientId}`}>
-                                            Marked for Deletion <button
-                                                onClick={(click) => handleUndoDelete(click, initialIngredient)}
-                                                key={`btn--dltIng1${initialIngredient.ingredientId}`}
-                                                className="btn--undoDelete">Undo</button>
-                                        </span>
-                                    </div>
-                                
-                                : 
-                                    <div className="addedIngredientRow" key={`initialIngDetails2--${initialIngredient.ingredientId}`}>
-                                        <span className="flex-column1" key={`matchedIng--${initialIngredient.ingredientId}`}>{matchedIngredient?.name}</span>
-                                        <span className="flex-column2" key={`addedQuant--${initialIngredient.ingredientId}`}>{initialIngredient.quantity} {initialIngredient.quantityUnit}</span>
-                                        <span className="flex-column3" key={`removeIngredient--${initialIngredient.ingredientId}`}>
-                                            <button data-id={initialIngredient.id}
-                                                onClick={(click) => handleDeleteExistingIngredient(click, initialIngredient)}
-                                                key={`btn--dltIng2${initialIngredient.ingredientId}`}
-                                                className="btn--removeItem">X</button>
-                                        </span>
-                                    </div>
-                                
-                        }
-                    </div>
-                })
-            }
+        {
+            !showCustom
+                ? <EditIngredientForm
+                initialIngredients={initialIngredients}
+                allIngredients={allIngredients}
+                markedForDeletion={markedForDeletion}
+                handleUndoDelete={handleUndoDelete}
+                handleDeleteExistingIngredient={handleDeleteExistingIngredient}
+                handleRemoveIngredient={handleRemoveIngredient}
+                ingredientsToPost={ingredientsToPost}
+                ingredientToAdd={ingredientToAdd}
+                updateIngredientToAdd={updateIngredientToAdd}
+                handleAddIngredient={handleAddIngredient}
+                setShowCustom={setShowCustom} />
 
-            {   /*New ingredients being added to recipe card by user*/
-                ingredientsToPost.length > 0
-                && ingredientsToPost.map(includedIngredient => {
-                    const matchedIngredient = allIngredients.find(
-                        ingredient => ingredient.id === includedIngredient.ingredientId
-                    )
-                    return <div className="addedIngredientRow addedIngredientRow--editForm" key={`addedIngDetails--${includedIngredient.ingredientId}`}>
-                        <span className="flex-column1" key={`matchedIng--${includedIngredient.ingredientId}`}>{matchedIngredient?.name}</span>
-                        <span className="flex-column2" key={`addedQuant--${includedIngredient.ingredientId}`}>{includedIngredient.quantity} {includedIngredient.quantityUnit}</span>
-                        <span className="flex-column3" key={`removeIngredient--${includedIngredient.ingredientId}`}>
-                            <button data-id={includedIngredient.ingredientId}
-                                onClick={(click) => handleRemoveIngredient(click, includedIngredient)}
-                                key={`btnremIng--${includedIngredient.ingredientId}`}
-                                className="btn--removeItem">X</button>
-                        </span>
-                    </div>
-                })
-            }
-        </div>
-        <section className="ingredientInputContainer">
-            <div className="form-group ingredientInputs">
-                <label>Choose Ingredient:
-                    <Select
-                        className="ingredient--control--select"
-                        id="ingredientChoices"
-                        options={allIngredients}
-                        onChange={(selectedOption) => {
-                            const copy = { ...ingredientToAdd }
-                            copy.ingredientId = parseInt(selectedOption.id)
-                            updateIngredientToAdd(copy)
-                            console.log("copyToAdd", copy)
-                        }}
-                        getOptionLabel={(option) => option.name}
-                        getOptionValue={(option) => option.id}
-                        placeholder="Select an Ingredient"
-                    />
-                </label>
-            </div>
-            <div className="form-group ingredientInputs">
-                <label htmlFor="ingredientQuantity_input">Quantity:</label>
-                <input
-                    required 
-                    type="number"
-                    className="ingredient--control"
-                    placeholder="Enter a quantity"
-                    id="ingredientQuantity_input"
-                    value={ingredientToAdd.quantity !== 0 ? ingredientToAdd.quantity : ""} // If value is is zero, change to empty string to display placeholder text by default instead of zero
-                    step="0.01" // Set the step attribute to control decimal precision
-                    onChange={
-                        (changeEvent) => {
-                            const copy = { ...ingredientToAdd }
-                            copy.quantity = changeEvent.target.value !== "" ? Math.round(parseFloat(changeEvent.target.value) * 100) / 100 : 0
-                            updateIngredientToAdd(copy) // Updating ingredient quantity with value of copy
-                        }
-                    } />
-            </div>
-            <div className="form-group ingredientInputs">
-                <label htmlFor="quantityType_input">Unit:</label>
-                <input
-                    required 
-                    type="text"
-                    className="ingredient--control"
-                    placeholder="Enter a unit of measurement"
-                    id="quantityType_input"
-                    value={ingredientToAdd.quantityUnit}
-                    onChange={
-                        (changeEvent) => {
-                            const copy = { ...ingredientToAdd }
-                            copy.quantityUnit = changeEvent.target.value
-                            updateIngredientToAdd(copy) // Updating recipe with value of copy
-                        }
-                    } />
-            </div>
-        </section>
-        <button
-            onClick={
-                (event) => {handleAddIngredient(event)}
-            }
-        >Add Ingredient</button>
+                : <CustomIngredient
+                allIngredients={allIngredients}
+                fetchList={fetchList}
+                setAllIngredients={setAllIngredients}
+                setShowCustom={setShowCustom} />
+        }
+        
     </>
 }
